+++ src/pages/post/Comment.tsx
@@ -8,19 +8,7 @@ import Badge from "./Badge";
 import { Link } from "react-router";
 import { Activity, useEffect, useRef, useState } from "react";
 import { useAuthStore } from "../../stores/authStore";
-
-function formatRelativeTime(dateString: string) {
-  const diff = Date.now() - new Date(dateString).getTime();
-  const seconds = Math.floor(diff / 1000);
-  const minutes = Math.floor(seconds / 60);
-  const hours = Math.floor(minutes / 60);
-  const days = Math.floor(hours / 24);
-  if (seconds < 60) return "Just now";
-  if (minutes < 60) return `${minutes}m ago`;
-  if (hours < 24) return `${hours}h ago`;
-  if (days < 7) return `${days}d ago`;
-  return new Date(dateString).toLocaleDateString();
-}
+import formatRelativeTime from "../../services/formatRelativeTime";
 
 export default function Comment({
   comment,
